+++ src/eca/features/tools.clj
@@ -230,7 +230,10 @@
                                                            :tool-call-id tool-call-id
                                                            :call-state-fn call-state-fn
                                                            :state-transition-fn state-transition-fn})
-                           (f.mcp/call-tool! tool-name arguments {:db db})))
+                           (f.mcp/call-tool! tool-name arguments {:db db
+                                                                  :db* db*
+                                                                  :config config
+                                                                  :metrics metrics})))
                        (tools.util/maybe-truncate-output config tool-call-id))]
         (logger/debug logger-tag "Tool call result: " result)
         (metrics/count-up! "tool-called" {:name full-name :error (:error result)} metrics)